Minor tweaks for effects

+++ DarniProject/Assets/Scripts/ApplyEffect.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public abstract class ApplyEffect : ScriptableObject
 {
-    public abstract void Apply(PlayerXP playerLevelManager, int value);
+    public abstract void Apply(PlayerXP playerLevelManager, float value);
 }
