+++ DarniProject/Assets/Scripts/ApplyEffects/IncreaseMovementSpeed.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 public class IncreaseMovementSpeed : ApplyEffect
 {
-    public override void Apply(PlayerXP player, int value)
+    public override void Apply(PlayerXP player, float value)
     {
         player.movementStats.topSpeed += value;
     }
