+++ DarniProject/Assets/Scripts/StatBonus.cs
@@ -3,17 +3,17 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Stat Bonus", menuName = "LevelUp/Stat Bonus")]
 public class StatBonus : LevelUpBonus
 {
-    public int[] valuesPerLevel;
+    public float[] valuesPerLevel;
     public ApplyEffect effect;
 
     public override void ApplyOrUnlock(PlayerXP playerLevelManager, int skillLevel)
     {
-        int value = GetValue(skillLevel);
+        float value = GetValue(skillLevel);
         effect.Apply(playerLevelManager, value);
     }
 
     // Temporary safety function to avoid out of bounds access
-    protected int GetValue(int level)
+    protected float GetValue(int level)
     {
         return valuesPerLevel[Mathf.Clamp(level, 0, valuesPerLevel.Length - 1)];
     }
